% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSVReader.R
\name{LA_ImportData.CSV}
\alias{LA_ImportData.CSV}
\title{R script for reaing CSV data}
\usage{
\method{LA_ImportData}{CSV}(filename, dec = ".", sep = ",", ...)
}
\arguments{
\item{dec}{, decimal separator}
}
\value{
data.frame
}
\description{
This function requires a table/list 
             'DiffTime', 'Value', 'Description', Wavelength'
}
\details{
LA_ImportData.CSV
}
\note{
todo -
}
\keyword{CSV}
\keyword{plate}
\keyword{readers,}
