% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importData.R
\name{LA_ImportData.default}
\alias{LA_ImportData.default}
\title{Importing Data from Microtiter Plate Readers - default method}
\usage{
\method{LA_ImportData}{default}(filename, ...)
}
\description{
This function is the default function
}
\details{
LA_ImportData.default
}

