% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plateReader.R
\name{LA_ImportData.omega}
\alias{LA_ImportData.omega}
\title{BMG omega - file reader}
\usage{
\method{LA_ImportData}{omega}(filename, coords = FALSE, ...)
}
\arguments{
\item{filename}{(string) - single varioskan outputfile of kinetic data}

\item{barcode="000000"}{(string) - for defining a new barcode}
}
\value{
data.frame
}
\description{
This function requires a table/list omega output file starting 
             from the keyword "Chromatic" with the following columns:
}
\details{
LA_ImportData.omega
}
\note{
todo - storing meta info
}
\keyword{BMG}
\keyword{omega}
\keyword{plate}
\keyword{readers,}

