% Generated by roxygen2 (4.0.2): do not edit by hand
\name{LA_ImportData}
\alias{LA_ImportData}
\title{Importing Data from Microtiter Plate Readers - generic method}
\usage{
LA_ImportData(filename = "default_file_name", data_path = "./",
  method = "SPabs", barcode = "0000", wavelength = 0, numSamples = 96,
  layout = FALSE, useDBlayout = FALSE, layoutDir = "./", PLC = FALSE,
  wellRadius = 0.34, liquidVolume = 0.2, ...)
}
\arguments{
\item{method}{(string) - one of the above described methods}

\item{filename/pattern}{(string)}

\item{path}{(string)}

\item{device}{(string) - one of the above mentioned devices}

\item{layout=TRUE}{(boolean) - automatically read plate layout file}

\item{layout_dir="./"}{(string) - plate layout file directory with leading /}

\item{use_db_layout=FALSE}{(boolean) - reading the layout from a database}

\item{PLC=FALSE}{(boolean) - automatically calculate pathlength correction
under the assumptions of a 96 round well SBS layout (r=034 cm)
and 200ul liquid volume}

\item{well_radius=0.34}{(double) well radius for pathlength calculation in cm}

\item{liquid_volume=0.2}{(double) liquid volume for pathlength calculation in uL}
}
\value{
data frame with reader data
}
\description{
This function is the main dispatcher function for reading several file formats into dataframes.
             Currently supported devices:
               BMG Readers (e.g. Omega ) - device='omega'
               Thermo Scientific (e.g. Varioskan) - device='varioskan'
             Currently supported formats (method):
               SPabs : Single point absorption measurement (Omega, Varioskan),
               SPabsMatr: Single point absorption measurement in Matrix Layout format (Omega),
               SPfl:   Single point fluorescense measurement (Varioskan),
               KINabs: Kinetic absorption measurement (Varioskan)
}
\details{
LA_ImportData
}
\note{
todo: debugging,  error handling, filename patterns
}
\examples{
kin_df <- importReaderData("_KINabs_245_265_", method="KINabs", device='varioskan', layout=FALSE, PLC=TRUE)
}
\keyword{plate}
\keyword{readers}

