% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plateReader.R
\name{LA_ImportData.tecanMagellanCSV}
\alias{LA_ImportData.tecanMagellanCSV}
\title{TECAN}
\usage{
\method{LA_ImportData}{tecanMagellanCSV}(filename, ...)
}
\arguments{
\item{filename}{(string) - TECAN Magellan MSExcel outputfile converted to CSV}
}
\value{
data.frame
}
\description{
This function requires a table/list TECAN Magellan CSV  output file starting 
             from the keyword "Raw data" with the following columns:
}
\details{
LA_ImportData.tecanMagellanCSV
}
\note{
todo - storing all meta info
}
\keyword{Thermo}
\keyword{Varioskan}
\keyword{plate}
\keyword{readers,}

