% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plateReader.R
\name{LA_ImportData.tecanMagellanXLSX}
\alias{LA_ImportData.tecanMagellanXLSX}
\title{TECAN Magellan MS Excel xlxs reading module}
\usage{
\method{LA_ImportData}{tecanMagellanXLSX}(filename, ...)
}
\arguments{
\item{filename}{(string) - TECAN Magellan MSExcel xlxs file name}
}
\value{
data.frame
}
\description{
This function requires a table/list TECAN Magellan XLSX output file starting 
             from the keyword "Photometric1" with the following columns:
}
\details{
LA_ImportData.tecanMagellanXLSX
}
\note{
todo - storing all meta info
}
\keyword{TECAN}
\keyword{plate}
\keyword{readers,}
