% Generated by roxygen2 (4.0.2): do not edit by hand
\name{LA_ImportData.varioskan}
\alias{LA_ImportData.varioskan}
\title{Importing Data from Microtiter Plate Readers}
\usage{
\method{LA_ImportData}{varioskan}(filename, coords = FALSE, ...)
}
\arguments{
\item{filename}{(string) - single varioskan outputfile of kinetic data}

\item{filename/pattern}{(string)}

\item{path}{(string)}

\item{method}{(string) - one of the above described methods}

\item{device}{(string) - one of the above mentioned devices}

\item{layout=TRUE}{(boolean) - automatically read plate layout file}

\item{layout_dir="./"}{(string) - plate layout file directory with leading /}

\item{use_db_layout=FALSE}{(boolean) - reading the layout from a database}

\item{PLC=FALSE}{(boolean) - automatically calculate pathlength correction
under the assumptions of a 96 round well SBS layout (r=034 cm)
and 200ul liquid volume}

\item{well_radius=0.34}{(double) well radius for pathlength calculation in cm}

\item{liquid_volume=0.2}{(double) liquid volume for pathlength calculation in uL}

\item{barcode="0000"}{(string) - for defining a new barcode}
}
\value{
data frame with reader data

data.frame
}
\description{
This function is the main dispatcher function for reading several file formats into dataframes.
             Currently supported devices:
               BMG Readers (e.g. Omega ) - device='omega'
               Thermo Scientific (e.g. Varioskan) - device='varioskan'
             Currently supported formats (method):
               SPabs : Single point absorption measurement (Omega, Varioskan),
               SPabsMatr: Single point absorption measurement in Matrix Layout format (Omega),
               SPfl:   Single point fluorescense measurement (Varioskan),
               KINabs: Kinetic absorption measurement (Varioskan)
}
\details{
LA_ImportData.varioskan
}
\note{
todo: debugging,  error handling, filename patterns

todo - adding number of measurements, Time as factors ??
}
\examples{
kin_df <- importReaderData("_KINabs_245_265_", method="KINabs", device='varioskan', layout=FALSE, PLC=TRUE)
}
\keyword{Thermo}
\keyword{Varioskan}
\keyword{plate}
\keyword{readers}
\keyword{readers,}

