% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{package}
\name{laraDataReader}
\alias{laraDataReader}
\alias{laraDataReader-package}
\title{Reading Various Laboratory Device File Formats and plate layouts}
\description{
Reading Various Laboratory Device File Formats and plate layouts
}
\details{
currently supported readers: Thermo Varioskan (and related readers), BMG Omega (and related BMG readers)
}
\note{
date: 150601

last modification:
}
\author{
mark doerr \email{mark.doerr@uni-greifswald.de},
  and \email{mark@ismeralda.org}
}

