% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plateReader.R
\name{omegaSPabsMatr}
\alias{omegaSPabsMatr}
\title{BMG Omega: Singlepoint absorption measurement with multiple wavelengths - old file formats (matrix)}
\usage{
omegaSPabsMatr(filename_list)
}
\arguments{
\item{filename_list}{(string) - list of filenames to import}
}
\value{
data.frame
}
\description{
This function can be used for matrix type BMG omega output data
}
\details{
omegaSPabsMatr
}
\note{
todo: temperature, shaking, num flashes
}
\keyword{BMG}
\keyword{omega}
\keyword{plate}
\keyword{readers,}

