% Generated by roxygen2 (4.0.2): do not edit by hand
\name{pathlengthCorrectionFactor}
\alias{pathlengthCorrectionFactor}
\title{A simple pathlenght correction factor for microtiterplaters}
\usage{
pathlengthCorrectionFactor(liquid_volume = 0.2, well_radius = 0.34)
}
\arguments{
\item{liquid_volume=0.2}{}

\item{well_radius=0.34}{(standard radius of well in 96 well MTPs)}
}
\description{
A simple pathlenght correction factor is generated, based on volume and radius of a cylinder
             More accurate pathlength correction is achieved by measureing the absorption of a solution
             with know concentration with different per well
             volumes and recording of a calibration curve.
}
\details{
pathlengthCorrectionFactor
}
\examples{
pl_factor <- pathlengthCorrectionFactor(liquid_volume=0.2, well_radius=0.34)
}
\keyword{correction}
\keyword{pathlength}
\keyword{plate}
\keyword{readers,}

