% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importData.R
\name{genWellNumbers}
\alias{genWellNumbers}
\title{Well Number Generator}
\usage{
genWellNumbers(padding = 2, nrows = 8, ncols = 12, repeats = 1,
  byRows = FALSE, asVector = FALSE)
}
\arguments{
\item{padding=2}{(overall number of digits, leading digits are filled with zeros)}

\item{nrows=8}{(number of rows on plate)}

\item{ncols=12}{(number of columns on plate)}
}
\description{
Generator for microtiter plate well coordinates with repeats
}
\details{
genWellNumbers
}
\examples{
 genWellNumbers(padding=1)
             >  A1, A2, ...., H12
           genWellNumbers(padding=3)
             > A001, A002, ..., H012
}
\keyword{microtiter}
\keyword{plate}
\keyword{plates}
\keyword{readers,}

