% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importData.R
\name{LA_ImportData}
\alias{LA_ImportData}
\title{Importing Data from Microtiter Plate Readers - generic method}
\usage{
LA_ImportData(filename = "", data_path = "./", method = "SPabs",
  defaultClass = "SP", barcode = "000000", wavelength = 0, wlStart = 0,
  wlEnd = 0, wlStepSize = 1, timeUnit = "secs", numSamples = 96,
  sep = ",", layoutGeometry = c(8, 12), layout = FALSE,
  useDBlayout = FALSE, layoutDir = "./", concUnit = "uM", PLC = FALSE,
  wellRadius = 0.34, liquidVolume = 0.2, ...)
}
\arguments{
\item{method}{(string) - one of the above described methods}

\item{layoutGeometry}{- container layout geometry}

\item{filename/pattern}{(string)}

\item{path}{(string)}

\item{device}{(string) - one of the above mentioned devices}

\item{layout=TRUE}{(boolean) - automatically read plate layout file}

\item{layoutDir="./"}{(string) - plate layout file directory with leading /}

\item{useDBlayout=FALSE}{(boolean) - reading the layout from a database}

\item{concUnit="uM"}{- conentration unit of layout file}

\item{PLC=FALSE}{(boolean) - automatically calculate pathlength correction 
under the assumptions of a 96 round well SBS layout (r=034 cm)
and 200ul liquid volume}

\item{wellRadius=0.34}{(double) well radius for pathlength calculation in cm}

\item{liquidVolume=0.2}{(double) liquid volume for pathlength calculation in uL}
}
\value{
data frame with reader data
}
\description{
This function is the main dispatcher function for reading several file formats into dataframes.
             Currently supported devices:
               BMG Readers (e.g. Omega ) - device='omega'
               Thermo Scientific (e.g. Varioskan) - device='varioskan'
               TECAN (e.g. ) - device='tecan'
             Currently supported formats (method):
               SPabs : Single point absorption measurement (Omega, Varioskan),
               SPabsMatr: Single point absorption measurement in Matrix Layout format (Omega),
               SPfl:   Single point fluorescense measurement (Varioskan),
               KINabs: Kinetic absorption measurement (Varioskan)
}
\details{
LA_ImportData
}
\note{
todo: debugging,  error handling, filename patterns
}
\examples{
          kin_df <- importReaderData("_KINabs_245_265_", method="KINabs", device='varioskan', layout=FALSE, PLC=TRUE)

}
\keyword{plate}
\keyword{readers}

