% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plateReader.R
\name{LA_ImportData.varioskan}
\alias{LA_ImportData.varioskan}
\title{Thermo Varioskan file reader}
\usage{
\method{LA_ImportData}{varioskan}(filename = "", deltaTime = 32,
  addTime = TRUE, coords = FALSE, read = FALSE, ...)
}
\arguments{
\item{filename}{(string) - single varioskan outputfile of kinetic data}

\item{barcode="000000"}{(string) - for defining a new barcode}
}
\value{
data.frame
}
\description{
This function requires a table/list Varioskan  output file starting 
             from the keyword "Photometric1" with the following columns:
             'Barcode', 'Well', 'Type', 'Description', 'SampleNo', 'Value', 'Time', 'Wavelength', 'Read'.
}
\details{
LA_ImportData.varioskan
}
\note{
todo - adding meta info (temp....), spectrum read: assigning start and end wavelength
}
\keyword{Thermo}
\keyword{Varioskan}
\keyword{plate}
\keyword{readers,}

