% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plateReader.R
\name{LA_ImportData.varioskan}
\alias{LA_ImportData.varioskan}
\title{Thermo Varioskan file reader}
\usage{
\method{LA_ImportData}{varioskan}(filename = "", deltaTime = 32,
  addTime = TRUE, coords = FALSE, read = FALSE, ...)
}
\arguments{
\item{deltaTime=32.0}{- default time difference between two reads}

\item{addTime=TRUE}{- adding DataTime info to output data frame}

\item{coords=FALSE}{- adding plate coords to output data frame (if provided by reader file)}

\item{read=FALSE}{- adding read info to output data frame (if provided by reader file)}

\item{barcode="000000"}{(string) - for defining a new barcode}
}
\value{
data.frame
}
\description{
This function requires a table/list Varioskan  output file starting 
             from the keyword "Photometric1" with the following columns:
             'Barcode', 'Well', 'Type', 'Description', 'SampleNo', 'Value', 'Time', 'Wavelength', 'Read'.
}
\details{
LA_ImportData.varioskan
}
\note{
todo - spectrum read: assigning start and end wavelength
}
\keyword{Thermo}
\keyword{Varioskan}
\keyword{plate}
\keyword{readers,}

