% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/laraDataReader-package.r
\docType{package}
\name{laraDataReader}
\alias{laraDataReader}
\alias{laraDataReader-package}
\title{Reading Various Laboratory Device File Formats and plate layouts}
\details{
currently supported readers: Thermo Varioskan (and related readers), BMG Omega (and related BMG readers)

Varioskan: Single point Absorbance in table format (SPabs), single point absorbance multiple wavelengths,

fluorescense
}
\note{
date: 150601

last modification:
}
\author{
mark doerr \email{mark.doerr@uni-greifswald.de},
  and \email{mark@ismeralda.org}
}
