% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layoutReader.R
\name{loadPlateLayout}
\alias{loadPlateLayout}
\title{Reading Plate Layout Information from Plate Layout csv File for Microtiter Plates.}
\usage{
loadPlateLayout(barcode = "000000", dir = "./", asList = TRUE)
}
\arguments{
\item{barcode="000000"}{- this is used for finding the layout file BC_plate_layout_*.csv}

\item{dir="./"}{(string) - directory of layout file with leading /}

\item{asList=TRUE}{(boolean) this returns not only the plate layout, but also the barcodes, description and plate geometry (= layout meta information)}
}
\value{
data.frame containing Wells, Type and Description, Substrates, Cofactors and Enzymes as columns
}
\description{
Reading plate layout information from plate layout csv file for any microtiter well plates.
             Plate geometry is defined in layout file. Default geometry 8x12.
             The function is looking for a file "0*"barcode"_plate_layout.*csv" in the given directory.
             Default substrate, cofactor and enzyme concentrations overwrite individual concentrations.
}
\details{
loadPlateLayout
}
\keyword{layout}
\keyword{microtiter}
\keyword{plate}
\keyword{readers}

